Move UI code behind to ViewModel
