+++ src/MainPage.xaml
@@ -2,9 +2,10 @@
     x:Class="FaceDetection.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:FaceDetection"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:i="using:Microsoft.Xaml.Interactivity"
+    xmlns:core="using:Microsoft.Xaml.Interactions.Core"
     mc:Ignorable="d">
 
     <Page.Resources>
@@ -25,28 +26,51 @@
             </Style>
         </Grid.Resources>
 
-        <CaptureElement Name="PreviewControl" Stretch="Uniform"/>
-        <Image Name="ImageControl" Stretch="Uniform"/>
+        <CaptureElement Name="PreviewControl" Stretch="Uniform">
+            <i:Interaction.Behaviors>
+                <core:EventTriggerBehavior EventName="Loaded">
+                    <core:InvokeCommandAction Command="{x:Bind ViewModel.PreviewControlLoaded}"
+                                              CommandParameter="{Binding ElementName=PreviewControl}"/>
+                </core:EventTriggerBehavior>
+            </i:Interaction.Behaviors>
+        </CaptureElement>
+        <Image Name="ImageControl" Stretch="Uniform">
+            <i:Interaction.Behaviors>
+                <core:EventTriggerBehavior EventName="Loaded">
+                    <core:InvokeCommandAction Command="{x:Bind ViewModel.ImageControlLoaded}"
+                                              CommandParameter="{Binding ElementName=ImageControl}"/>
+                </core:EventTriggerBehavior>
+            </i:Interaction.Behaviors>
+        </Image>
         <Canvas>
-            <Canvas Name="FacesCanvas"/>
+            <Canvas Name="FacesCanvas">
+                <i:Interaction.Behaviors>
+                    <core:EventTriggerBehavior EventName="Loaded">
+                        <core:InvokeCommandAction Command="{x:Bind ViewModel.FacesCanvasLoaded}"
+                                                  CommandParameter="{Binding ElementName=FacesCanvas}"/>
+                    </core:EventTriggerBehavior>
+                </i:Interaction.Behaviors>
+            </Canvas>
         </Canvas>
 
         <StackPanel HorizontalAlignment="Right" VerticalAlignment="Center">
-            <Button Name="PhotoButton" Click="PhotoButton_Click">
+            <Button Name="PhotoButton"
+                    Command="{x:Bind ViewModel.PhotoButtonClicked}">
                 <Viewbox>
                     <SymbolIcon Symbol="Pictures"/>
                 </Viewbox>
             </Button>
-            <Button Name="CameraButton" Click="CameraButton_Click">
+            <Button Name="CameraButton" 
+                    Command="{x:Bind ViewModel.CameraButtonClicked}">
                 <Viewbox>
                     <SymbolIcon Symbol="Camera"/>
                 </Viewbox>
             </Button>
-            <Button Name="FaceDetectionButton" Click="FaceDetectionButton_Click">
+            <Button Name="FaceDetectionButton"
+                    Command="{x:Bind ViewModel.FaceDetectionButtonClicked}">
                 <Viewbox>
                     <Grid>
-                        <SymbolIcon Name="FaceDetectionDisabledIcon" Symbol="Contact" Visibility="Visible"/>
-                        <SymbolIcon Name="FaceDetectionEnabledIcon" Symbol="Contact2" Visibility="Collapsed"/>
+                        <SymbolIcon Name="FaceDetectionDisabledIcon" Symbol="Contact"/>
                     </Grid>
                 </Viewbox>
             </Button>
