+++ src/Models/FrameModel.cs
@@ -8,7 +8,7 @@ using Windows.Graphics.Imaging;
 
 namespace FaceDetection.Models
 {
-    public class FrameModel : BaseNotifyPropertyChanged
+    public class FrameModel : SingletonNotifyPropertyChanged<FrameModel>
     {
         private SoftwareBitmap _softwareBitmap;
         public SoftwareBitmap SoftwareBitmap
