+++ src/ViewModels/MainPageViewModel.cs
@@ -8,54 +8,81 @@ using System.Text;
 using System.Threading.Tasks;
 using Windows.Graphics.Imaging;
 using Windows.Storage;
-using Windows.Storage.Streams;
 using FaceDetection.FaceDetector;
 using System.Diagnostics;
-using Windows.Media.Capture;
 using Windows.Media.Capture.Frames;
-using Windows.Devices.Enumeration;
-using Windows.Media.MediaProperties;
 using System.Threading;
-using System.IO;
 using FaceDetection.DistanceEstimator;
 using System.ComponentModel;
+using System.Windows.Input;
+using Windows.UI.Xaml.Controls;
+using Windows.UI.Xaml.Media.Imaging;
+using Windows.UI.Core;
+using Windows.UI.Xaml.Shapes;
+using Windows.UI.Xaml.Media;
+using Windows.UI;
+using Windows.Foundation;
+using Microsoft.Toolkit.Uwp.Helpers;
+using Windows.Storage.Pickers;
 
 namespace FaceDetection.ViewModels
 {
     public class MainPageViewModel : BaseNotifyPropertyChanged
     {
-        public FrameModel FrameModel { get; } = new FrameModel();
+        public FrameModel FrameModel { get; } = FrameModel.Instance;
         public CameraControl CameraControl { get; } = new CameraControl();
         private int _processingFlag;
 
         private IFaceDetector _faceDetector;
         private bool _isFaceDetectionEnabled = false;
-        // TODO: Create Config to load modelFileName
-        private string _modelFileName = "version-RFB-320.onnx";
-        
-        public event FaceDetectedEventHandler FaceDetected
-        {
-            add
-            {
-                lock (_faceDetector)
-                {
-                    _faceDetector.FaceDetected += value;
-                }
-            }
-            remove
-            {
-                lock (_faceDetector)
-                {
-                    _faceDetector.FaceDetected -= value;
-                }
-            }
-        }
         public bool IsFaceDetectionEnabled
         {
             get => _isFaceDetectionEnabled;
             set
             {
                 SetProperty(ref _isFaceDetectionEnabled, value);
+            }
+        }
+        // TODO: Create Config to load modelFileName
+        private string _modelFileName = "version-RFB-320.onnx";
+
+        public ICommand ImageControlLoaded { get; set; }
+        public ICommand PreviewControlLoaded { get; set; }
+        public ICommand FacesCanvasLoaded { get; set; }
+
+        public ICommand PhotoButtonClicked { get; set; }
+        public ICommand CameraButtonClicked { get; set; }
+        public ICommand FaceDetectionButtonClicked { get; set; }
+
+        private Image _imageControl;
+        public Image ImageControl
+        {
+            get => _imageControl;
+            set
+            {
+                SetProperty(ref _imageControl, value);
+            }
+        }
+
+        private CaptureElement _previewControl;
+        public CaptureElement PreviewControl
+        {
+            get => _previewControl;
+            set
+            {
+                SetProperty(ref _previewControl, value);
+            }
+        }
+
+        private Canvas _facesCanvas;
+
+        private int _previewWidth;
+        public int PreviewWidth
+        {
+            get => _previewWidth;
+            set
+            {
+                SetProperty(ref _previewWidth, value);
             }
         }
 
@@ -65,6 +92,79 @@ namespace FaceDetection.ViewModels
         public MainPageViewModel()
         {
             SubscribeEvents();
+
+            ImageControlLoaded = new DelegateCommand<Image>(imageControl =>
+            {
+                _imageControl = imageControl;
+            });
+            PreviewControlLoaded = new DelegateCommand<CaptureElement>(previewControl =>
+            {
+                _previewControl = previewControl;
+            });
+            FacesCanvasLoaded = new DelegateCommand<Canvas>(facesCanvas =>
+            {
+                _facesCanvas = facesCanvas;
+            });
+
+            PhotoButtonClicked = new DelegateCommand(async () =>
+            {
+                var picker = new FileOpenPicker();
+                picker.ViewMode = PickerViewMode.Thumbnail;
+                picker.SuggestedStartLocation = PickerLocationId.PicturesLibrary;
+                picker.FileTypeFilter.Add(".jpg");
+                picker.FileTypeFilter.Add(".jpeg");
+                picker.FileTypeFilter.Add(".png");
+                picker.FileTypeFilter.Add("*");
+                StorageFile file = await picker.PickSingleFileAsync();
+                if (file != null)
+                {
+                    using (var fileStream = await file.OpenAsync(FileAccessMode.Read))
+                    {
+                        await ClearFacesCanvas();
+                        if (CameraControl.IsPreviewing)
+                        {
+                            await StopPreviewAsync();
+                        }
+                        BitmapImage bmp = new BitmapImage();
+                        bmp.DecodePixelHeight = (int)ImageControl.Height;
+                        bmp.DecodePixelWidth = (int)ImageControl.Width;
+                        await bmp.SetSourceAsync(fileStream);
+                        ImageControl.Source = bmp;
+                        
+                        BitmapDecoder decoder = await BitmapDecoder.CreateAsync(fileStream);
+                        SoftwareBitmap softwareBitmap = await decoder.GetSoftwareBitmapAsync(BitmapPixelFormat.Bgra8, BitmapAlphaMode.Ignore);
+                        FrameModel.SoftwareBitmap = softwareBitmap;
+                    }
+                }
+            });
+            CameraButtonClicked = new DelegateCommand(async () =>
+            {
+                if (!CameraControl.IsPreviewing)
+                {
+                    ImageControl.Source = null;
+                    await InitCameraAsync();
+                    PreviewControl.Source = CameraControl.MediaCapture;
+                    await StartPreviewAsync();
+                }
+                else
+                {
+                    PreviewControl.Source = null;
+                    await StopPreviewAsync();
+                }
+                await ClearFacesCanvas();
+            });
+            FaceDetectionButtonClicked = new DelegateCommand(async () =>
+            {
+                IsFaceDetectionEnabled = !IsFaceDetectionEnabled;
+                if (IsFaceDetectionEnabled)
+                {
+                    _faceDetector.FaceDetected += _faceDetector_FaceDetected;
+                } else
+                {
+                    _faceDetector.FaceDetected -= _faceDetector_FaceDetected;
+                }
+                await ClearFacesCanvas();
+            });
         }
 
         public void OnFrameArrived(MediaFrameReader sender, MediaFrameArrivedEventArgs args)
@@ -112,6 +212,104 @@ namespace FaceDetection.ViewModels
         public float EstimateDistance(FaceBoundingBox bb)
         {
             return _distanceEstimator.ComputeDistance(bb);
+        }
+
+        private async void _faceDetector_FaceDetected(object sender, IReadOnlyList<FaceBoundingBox> faceBoundingBoxes, System.Drawing.Size originalSize)
+        {
+            await DispatcherHelper.ExecuteOnUIThreadAsync(() => HighlightDetectedFaces(faceBoundingBoxes, originalSize));
+        }
+
+        private void HighlightDetectedFaces(IReadOnlyList<FaceBoundingBox> faces, System.Drawing.Size originalSize)
+        {
+            _facesCanvas.Children.Clear();
+            for (int i = 0; i < faces.Count; ++i)
+            {
+                var bb = ScaleBoundingBox(faces[i], originalSize);
+                Rectangle faceBB = ConvertPreviewToUiRectangle(bb, originalSize);
+                faceBB.StrokeThickness = 2;
+                faceBB.Stroke = new SolidColorBrush(Colors.LimeGreen);
+                _facesCanvas.Children.Add(faceBB);
+
+                TextBlock txtBlk = new TextBlock();
+                txtBlk.Text = EstimateDistance(faces[i]).ToString("n0") + " cm";
+                txtBlk.Foreground = new SolidColorBrush(Colors.LimeGreen);
+                Canvas.SetLeft(txtBlk, Canvas.GetLeft(faceBB));
+                Canvas.SetTop(txtBlk, Canvas.GetTop(faceBB));
+                _facesCanvas.Children.Add(txtBlk);
+            }
+        }
+
+        private Rectangle ConvertPreviewToUiRectangle(FaceBoundingBox faceBox, System.Drawing.Size actualContentSize)
+        {
+            var result = new Rectangle();
+            double streamWidth = actualContentSize.Width;
+            double streamHeight = actualContentSize.Height;
+
+            //If there is no available information about the preview, return an empty rectangle, as re - scaling to the screen coordinates will be impossible
+            //  Similarly, if any of the dimensions is zero(which would only happen in an error case) return an empty rectangle
+            if (streamWidth == 0 || streamHeight == 0) return result;
+
+
+            //Get the rectangle that is occupied by the actual video feed
+            var previewInUI = GetDisplayRectInControl(actualContentSize, PreviewControl);
+            var scaleWidth = previewInUI.Width / streamWidth;
+            var scaleHeight = previewInUI.Height / streamHeight;
+
+            //Scale the width and height from preview stream coordinates to window coordinates
+            result.Width = faceBox.Width * scaleWidth;
+            result.Height = faceBox.Height * scaleHeight;
+
+            // Scale the X and Y coordinates from preview stream coordinates to window coordinates
+            var x = previewInUI.X + faceBox.X0 * scaleWidth;
+            var y = previewInUI.Y + faceBox.Y0 * scaleHeight;
+
+            Canvas.SetLeft(result, x);
+            Canvas.SetTop(result, y);
+
+            return result;
+        }
+
+        private Rect GetDisplayRectInControl(System.Drawing.Size actualContentSize, CaptureElement previewControl)
+        {
+            var result = new Rect();
+            // In case this function is called before everything is initialized correctly, return an empty result
+            if (previewControl == null || previewControl.ActualHeight < 1 || previewControl.ActualWidth < 1 ||
+                actualContentSize.Height == 0 || actualContentSize.Width == 0)
+            {
+                return result;
+            }
+
+            var streamWidth = actualContentSize.Width;
+            var streamHeight = actualContentSize.Height;
+
+            // Start by assuming the preview display area in the control spans the entire width and height both (this is corrected in the next if for the necessary dimension)
+            result.Width = previewControl.ActualWidth;
+            result.Height = previewControl.ActualHeight;
+
+            // If UI is "wider" than preview, letterboxing will be on the sides
+            if ((previewControl.ActualWidth / previewControl.ActualHeight > streamWidth / (double)streamHeight))
+            {
+                var scale = previewControl.ActualHeight / streamHeight;
+                var scaledWidth = streamWidth * scale;
+
+                result.X = (previewControl.ActualWidth - scaledWidth) / 2.0;
+                result.Width = scaledWidth;
+            }
+            else // Preview stream is "wider" than UI, so letterboxing will be on the top+bottom
+            {
+                var scale = previewControl.ActualWidth / streamWidth;
+                var scaledHeight = streamHeight * scale;
+
+                result.Y = (previewControl.ActualHeight - scaledHeight) / 2.0;
+                result.Height = scaledHeight;
+            }
+
+            return result;
+        }
+
+        private async Task ClearFacesCanvas()
+        {
+            await DispatcherHelper.ExecuteOnUIThreadAsync(() => _facesCanvas.Children.Clear());
         }
 
         public async Task LoadModelAsync()
